Move AWS exporting to proper subfolder
