+++ monkey/monkey_island/cc/services/report.py
@@ -11,7 +11,7 @@ from six import text_type
 
 from monkey_island.cc.database import mongo
 from monkey_island.cc.models import Monkey
-from monkey_island.cc.report_exporter_manager import ReportExporterManager
+from cc.services.reporting.report_exporter_manager import ReportExporterManager
 from monkey_island.cc.services.config import ConfigService
 from monkey_island.cc.services.edge import EdgeService
 from monkey_island.cc.services.node import NodeService
