+++ monkey/monkey_island/cc/services/reporting/exporter_init.py
@@ -1,7 +1,7 @@
 import logging
 
-from monkey_island.cc.report_exporter_manager import ReportExporterManager
-from monkey_island.cc.resources.aws_exporter import AWSExporter
+from cc.services.reporting.report_exporter_manager import ReportExporterManager
+from cc.services.reporting.aws_exporter import AWSExporter
 from monkey_island.cc.services.remote_run_aws import RemoteRunAwsService
 
 logger = logging.getLogger(__name__)
